macOS support
This breaks EVERYTHING right now, sorry (it does run on my M2 Air though!)

+++ Core/Tests/Utilities.cc
@@ -15,7 +15,7 @@ void CellEntry(Reference<String> parameterString) {
 
     CELL_ASSERT(Byteswap((uint16_t)0xf0ff) == 0xfff0);
     CELL_ASSERT(Byteswap(0xf0fffeff) == 0xfffefff0);
-    CELL_ASSERT(Byteswap(0xf0fffeff00fffefe) == 0xfefeff00fffefff0);
+    CELL_ASSERT(Byteswap((uint64_t)0xf0fffeff00fffefe) == 0xfefeff00fffefff0);
 
     CELL_ASSERT(Clamp(2, 1, 3) == 2);
     CELL_ASSERT(Clamp(0, 1, 3) == 1);
